+++ src/pages/Section3/PartIntro/index.js
@@ -1,5 +1,4 @@
-import React, {useState} from 'react';
-import Heading from '../../../components/Heading';
+import React from 'react';
 import Oval from '../../../components/Oval';
 import styles from './index.module.css';
 import cx from 'classnames';
@@ -26,14 +25,14 @@ const PartIntro = ({
         <p className={cx(styles.text, styles.paragraph)}>{intro.post}</p>
       </div>
       <div className={styles.categoryIntro}>
-        {content.map(tech => (
+        {content?.map(tech => (
           <div className={styles.techColumn} key={tech.title}>
             <Oval className={cx(styles.oval, styles.white)}>
               <img src={tech.icon} className={styles.image} />
             </Oval>
             <h2 className={styles.sectionHeading}>{tech.title}</h2>
             <ul>
-              {tech.list.map(v => (
+              {tech?.list?.map(v => (
                 <li
                   className={cx(styles.contentItem, {
                     [styles.contentItemCreated]: linksMapping.find(mapping => mapping.label === v.name)
